requirements_barebones.txt
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
pyserial==3.4
python-engineio==4.0.0
python-socketio==5.0.4
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vosk==0.3.21
Werkzeug==1.0.1
xArm-Python-SDK==1.6.5
